Questions/requirement.txt
asgiref==3.2.5
asn1crypto==1.3.0
astroid==2.3.3
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
confusable-homoglyphs==3.2.0
cryptography==2.8
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.41.0
django-core==1.4.1
django-crispy-forms==1.9.0
django-registration==3.1
django-rest-auth==0.9.5
django-webpack-loader==0.7.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mysqlclient==1.4.6
numpy==1.18.1
oauthlib==3.1.0
Pillow==7.1.1
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
PySocks==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
